[REFACT] 인기 상품 조회 controller 메서드를 Product → OrderStatistics 도메인으로 이동
